(#173) Refactoring
