+++ src/Lanceur.Core/Managers/IExecutionManager.cs
@@ -1,4 +1,4 @@
-﻿using Lanceur.Core.Models;
+﻿using Lanceur.Core.Requests;
 
 namespace Lanceur.Core.Managers
 {
@@ -7,52 +7,6 @@ namespace Lanceur.Core.Managers
         #region Methods
 
         Task<ExecutionResponse> ExecuteAsync(ExecutionRequest request);
-
-        #endregion Methods
-    }
-
-    public class ExecutionRequest
-    {
-        #region Properties
-        public string Query { get; set; }
-        public bool ExecuteWithPrivilege { get; set; }
-        public QueryResult QueryResult { get; set; }
-
-        #endregion Properties
-    }
-
-    public class ExecutionResponse
-    {
-        #region Properties
-
-        public static ExecutionResponse EmptyResult => new()
-        {
-            HasResult = true,
-            Results = new List<QueryResult>()
-        };
-
-        public static ExecutionResponse NoResult => new()
-        {
-            HasResult = false,
-            Results = new List<QueryResult>()
-        };
-
-        public bool HasResult { get; set; }
-
-        public IEnumerable<QueryResult> Results { get; set; }
-
-        #endregion Properties
-
-        #region Methods
-
-        public static ExecutionResponse FromResults(IEnumerable<QueryResult> results)
-        {
-            return new ExecutionResponse
-            {
-                Results = results,
-                HasResult = results.Any(),
-            };
-        }
 
         #endregion Methods
     }
