+++ src/Lanceur.Core/Requests/AliasExecutionRequest.cs
@@ -1,0 +1,19 @@
+﻿namespace Lanceur.Core.Requests
+{
+    public class AliasExecutionRequest
+    {
+        #region Properties
+
+        public string Query { get; set; }
+
+        public bool RunAsAdmin { get; set; }
+
+        #endregion Properties
+
+        #region Methods
+
+        public static implicit operator AliasExecutionRequest(string query) => new() { Query = query };
+
+        #endregion Methods
+    }
+}
