+++ src/Lanceur.Core/Requests/NavigationResponse.cs
@@ -1,0 +1,14 @@
+﻿using Lanceur.Core.Models;
+
+namespace Lanceur.Core.Requests
+{
+    public class NavigationResponse
+    {
+        #region Properties
+
+        public QueryResult CurrentAlias { get; set; }
+        public string Query { get; set; }
+
+        #endregion Properties
+    }
+}
