+++ src/Lanceur.Tests/Functional/ExecutionManagerShould.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using Lanceur.Core.Managers;
 using Lanceur.Core.Models;
+using Lanceur.Core.Requests;
 using Lanceur.Core.Services;
 using Lanceur.Infra.Managers;
 using Lanceur.Macros;
