+++ src/Lanceur.Tests/ViewModels/MainViewModelShould.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using Lanceur.Core.Managers;
 using Lanceur.Core.Models;
+using Lanceur.Core.Requests;
 using Lanceur.Core.Services;
 using Lanceur.Core.Stores;
 using Lanceur.Infra.Managers;
