+++ src/Lanceur/Utils/UiMixin.cs
@@ -1,17 +1,14 @@
 ﻿using Lanceur.Core.Models;
+using Lanceur.Core.Requests;
 using System.Windows;
 using System.Windows.Controls;
-using static Lanceur.Views.MainViewModel;
 
 namespace Lanceur.Utils
 {
     internal static class UiMixin
     {
         #region Methods
-        public static string GetTextFromTextbox(this object source)
-        {
-            return source is TextBox tb ? tb.Text : string.Empty;
-        }
+
         public static AliasExecutionRequest GetQueryFromDataContext(this object source)
         {
             var param = source is FrameworkElement e && e?.DataContext is QueryResult result
@@ -22,6 +19,11 @@ namespace Lanceur.Utils
                 Query = param,
                 RunAsAdmin = false,
             };
+        }
+
+        public static string GetTextFromTextbox(this object source)
+        {
+            return source is TextBox tb ? tb.Text : string.Empty;
         }
 
         #endregion Methods
